Resolve circular import in CampAway, extract MoonSign
