+++ src/index.ts
@@ -18,6 +18,7 @@ export {
   skillModifiers,
 } from "./modifierTypes";
 export * from "./mood";
+export * from "./moonSign";
 export * from "./propertyTyping";
 export * from "./resources";
 export * from "./since";
